+++ cuda/src/kernels/nn/softmax.rs
@@ -56,8 +56,8 @@ impl Softmax {
         let mut launch_args = TractLaunchArgs::new(stream, &func);
         launch_args.set_view(&i_view);
         launch_args.set_view(&o_view);
-        launch_args.set_slice::<i64>(&shape_nd3);
-        launch_args.set_slice::<i64>(&strides_nd3);
+        launch_args.set_slice::<i32>(&shape_nd3);
+        launch_args.set_slice::<i32>(&strides_nd3);
 
         let cfg = LaunchConfig {
             grid_dim: ((shape_nd3[0] * shape_nd3[2]) as _, 1, 1),
